Don't show dialog if tag.svg can't be loaded.
gdk-pixbuf support for svg seems to be broken on my system.
Because of that Gdk::Pixbuf::create_from_file throws an Gdk::PixbufError
(UNKNOWN_TYPE), which results in an exceptionDialog to be shown.
By catching this exception early, the only negative effect of a broken
system is a missing icon.

closes #1